Move s3 message utils out to a generic place
