+++ src/lambda/commands/importLambda.ts
@@ -9,7 +9,7 @@ import * as _ from 'lodash'
 import * as path from 'path'
 import * as vscode from 'vscode'
 import { LambdaFunctionNode } from '../explorer/lambdaFunctionNode'
-import { showConfirmationMessage } from '../../s3/util/messages'
+import { showConfirmationMessage } from '../../shared/utilities/messages'
 import { LaunchConfiguration, getReferencedHandlerPaths } from '../../shared/debug/launchConfiguration'
 import { ext } from '../../shared/extensionGlobals'
 import { makeTemporaryToolkitFolder, fileExists, tryRemoveFolder } from '../../shared/filesystemUtilities'
