+++ src/lambda/commands/uploadLambda.ts
@@ -11,7 +11,7 @@ const localize = nls.loadMessageBundle()
 import * as AdmZip from 'adm-zip'
 import * as fs from 'fs'
 import * as path from 'path'
-import { showConfirmationMessage } from '../../s3/util/messages'
+import { showConfirmationMessage } from '../../shared/utilities/messages'
 import { ext } from '../../shared/extensionGlobals'
 import { fileExists, makeTemporaryToolkitFolder, tryRemoveFolder } from '../../shared/filesystemUtilities'
 import * as localizedText from '../../shared/localizedText'
