+++ src/s3/commands/createBucket.ts
@@ -9,7 +9,7 @@ import { localize } from '../../shared/utilities/vsCodeUtils'
 import { Commands } from '../../shared/vscode/commands'
 import { Window } from '../../shared/vscode/window'
 import { S3Node } from '../explorer/s3Nodes'
-import { showErrorWithLogs } from '../util/messages'
+import { showErrorWithLogs } from '../../shared/utilities/messages'
 import { validateBucketName } from '../util/validateBucketName'
 
 /**
