+++ src/s3/commands/downloadFileAs.ts
@@ -11,9 +11,10 @@ import { getLogger } from '../../shared/logger'
 import * as telemetry from '../../shared/telemetry/telemetry'
 import { Window } from '../../shared/vscode/window'
 import { S3FileNode } from '../explorer/s3FileNode'
-import { readablePath, showErrorWithLogs, showOutputMessage } from '../util/messages'
+import { readablePath } from '../util/util'
 import { progressReporter } from '../util/progressReporter'
 import { localize } from '../../shared/utilities/vsCodeUtils'
+import { showErrorWithLogs, showOutputMessage } from '../../shared/utilities/messages'
 
 /**
  * Downloads a file represented by the given node.
